Обновлены модели. Добавлены функции для Админа. Разделена логика для Репозиториев

+++ KuzinShop/Migrations/ApplicationContextModelSnapshot.cs
@@ -24,11 +24,11 @@ namespace KuzinShop.Migrations
 
             modelBuilder.Entity("KuzinShop.Models.AttributeModel", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"), 1L, 1);
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"), 1L, 1);
 
                     b.Property<string>("DataType")
                         .HasColumnType("nvarchar(max)");
@@ -47,17 +47,17 @@ namespace KuzinShop.Migrations
 
             modelBuilder.Entity("KuzinShop.Models.CategoryAttributeModel", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"), 1L, 1);
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"), 1L, 1);
 
-                    b.Property<int>("AttributeId")
-                        .HasColumnType("int");
+                    b.Property<long>("AttributeId")
+                        .HasColumnType("bigint");
 
-                    b.Property<int>("CategoryId")
-                        .HasColumnType("int");
+                    b.Property<long>("CategoryId")
+                        .HasColumnType("bigint");
 
                     b.HasKey("Id");
 
@@ -70,11 +70,11 @@ namespace KuzinShop.Migrations
 
             modelBuilder.Entity("KuzinShop.Models.CategoryModel", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"), 1L, 1);
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"), 1L, 1);
 
                     b.Property<string>("Name")
                         .IsRequired()
@@ -87,14 +87,14 @@ namespace KuzinShop.Migrations
 
             modelBuilder.Entity("KuzinShop.Models.ProductAttributeModel", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"), 1L, 1);
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"), 1L, 1);
 
-                    b.Property<int>("AttributeId")
-                        .HasColumnType("int");
+                    b.Property<long>("AttributeId")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime?>("DateValue")
                         .HasColumnType("datetime2");
@@ -102,8 +102,8 @@ namespace KuzinShop.Migrations
                     b.Property<int?>("IntegerValue")
                         .HasColumnType("int");
 
-                    b.Property<int>("ProductId")
-                        .HasColumnType("int");
+                    b.Property<long>("ProductId")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("StringValue")
                         .HasColumnType("nvarchar(max)");
@@ -119,14 +119,14 @@ namespace KuzinShop.Migrations
 
             modelBuilder.Entity("KuzinShop.Models.ProductModel", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("int");
+                        .HasColumnType("bigint");
 
-                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<int>("Id"), 1L, 1);
+                    SqlServerPropertyBuilderExtensions.UseIdentityColumn(b.Property<long>("Id"), 1L, 1);
 
-                    b.Property<int>("CategoryId")
-                        .HasColumnType("int");
+                    b.Property<long>("CategoryId")
+                        .HasColumnType("bigint");
 
                     b.Property<DateTime>("Date")
                         .HasColumnType("datetime2");
