+++ KuzinShop/Models/DTO/CreateProductDTO.cs
@@ -2,7 +2,7 @@
 {
     public class CreateProductDTO
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public string Publisher { get; set; }
@@ -10,7 +10,7 @@
         public string Image { get; set; }
         public int Price { get; set; }
 
-        public int CategoryId { get; set; }
+        public long CategoryId { get; set; }
 
         public List<CategoryModel> Categories { get; set; } = new();
         public List<AttributeDTO> Attributes { get; set; } = new();
@@ -18,7 +18,7 @@
 
     public class AttributeDTO
     {
-        public int AttributeId { get; set; }
+        public long AttributeId { get; set; }
         public string Name { get; set; } // Название атрибута
         public string? StringValue { get; set; }
         public int? IntegerValue { get; set; }
